Chart's target attribute has just subset of attributes in intersection
JIRA: TNT-87
Even that charts currently support only drill from data point which offers full intersection it could lead to the invalid drill to uri definitions in future once the drill from axis label will be supported. From this reason we decided to provide only intersection attributes working also with drill from labels - attributes from same dimension as origin and limited to the origin itself and his parents. The same is already done for pivot table
